Add more benchmarks
